[Camera] Add new internal API - extra preview event (#3011)
* [Camera] Add new internal API - extra preview event
