+++ src/Tizen.Multimedia.Camera/Camera/CameraCapturingEventArgs.cs
