+++ src/Tizen.Multimedia.Camera/Camera/CameraDeviceStateChangedEventArgs.cs
