+++ src/Tizen.Multimedia.Camera/Camera/CameraErrorOccurredEventArgs.cs
