+++ src/Tizen.Multimedia.Camera/Camera/CameraFocusStateChangedEventArgs.cs
