+++ src/Tizen.Multimedia.Camera/Camera/CameraInterruptStartedEventArgs.cs
