+++ src/Tizen.Multimedia.Camera/Camera/CameraInterruptedEventArgs.cs
