+++ src/Tizen.Multimedia.Camera/Camera/CameraStateChangedEventArgs.cs
