+++ src/Tizen.Multimedia.Camera/Camera/DoublePlane.cs
