+++ src/Tizen.Multimedia.Camera/Camera/FaceDetectedEventArgs.cs
