+++ src/Tizen.Multimedia.Camera/Camera/FaceDetectionData.cs
