+++ src/Tizen.Multimedia.Camera/Camera/HdrCaptureProgressEventArgs.cs
