+++ src/Tizen.Multimedia.Camera/Camera/MediaPacketPreviewEventArgs.cs
