+++ src/Tizen.Multimedia.Camera/Camera/SinglePlane.cs
