+++ src/Tizen.Multimedia.Camera/Camera/StillImage.cs
