+++ src/Tizen.Multimedia.Camera/Camera/TriplePlane.cs
